+++ .dart_tool/package_config.json
@@ -440,7 +440,7 @@
       "languageVersion": "3.0"
     }
   ],
-  "generated": "2023-11-08T16:01:32.469559Z",
+  "generated": "2023-11-08T18:48:55.450572Z",
   "generator": "pub",
   "generatorVersion": "3.0.5"
 }
