+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:blurting/mainApp.dart';
 import 'package:provider/provider.dart';
-import 'package:blurting/Static/provider.dart';
+import 'package:blurting/Utils/provider.dart';
 import 'package:socket_io_client/socket_io_client.dart' as IO;
 
 void main() {
