+++ lib/model/QnAModel.dart
@@ -1,0 +1,7 @@
+class GetAnswer{
+  String answer;
+
+  GetAnswer({required this.answer});
+
+  
+}
