+++ lib/service/QnASurvice.dart
@@ -1,0 +1,6 @@
+import 'package:http/http.dart' as http;
+import 'package:flutter/material.dart';
+
+class fetchAnswer extends ChangeNotifier{
+  
+}
